core: bl::time::seconds truncated to 32 bits on Windows
time::seconds() cast its argument to long before handing it to Boost, and long
is 32 bits on Windows (LLP64) where it is 64 on Linux. Any duration above
LONG_MAX seconds therefore wrapped NEGATIVE on Windows while being correct
everywhere else. Measured:

    clamped (int64)   = 3155760000
    static_cast<long> = -1139207296

The milliseconds() and microseconds() wrappers immediately below it already
cast to std::int64_t; seconds() was the odd one out. Boost's posix_time::seconds
takes any integral type, so passing int64 through needs nothing else.

Found through CookieJar_ExpiryAndMaxAgePrecedenceTests, which failed on every
Windows combination of the matrix and passes on Linux. A Max-Age larger than a
ptime can hold is clamped to MAX_MAX_AGE_SECONDS (3155760000, a hundred years)
precisely so that the expiry cannot overflow into the past and delete the
cookie - and the clamped value is itself above LONG_MAX, so the cast undid the
clamp and produced the exact outcome the clamp exists to prevent. The comment
above it warned about that outcome.

This is not a test-only defect. Any Windows caller of time::seconds() with a
duration above 68 years was affected, and a cookie with a long Max-Age was
discarded the moment it was stored.

CookieJar's own redundant static_cast< long > goes too, since it would truncate
before the fixed wrapper ever saw the value.

+++ src/include/baselib/httpclient/CookieJar.h
@@ -827,7 +827,7 @@ namespace bl
                     cookie.expiryTime = ( maxAge <= 0 )
                         ? time::ptime( time::neg_infin )
                         : when + time::seconds(
-                            static_cast< long >( std::min< std::int64_t >( maxAge, MAX_MAX_AGE_SECONDS ) )
+                            std::min< std::int64_t >( maxAge, MAX_MAX_AGE_SECONDS )
                             );
                 }
                 else if( haveExpires )
